+++ packages/react-container/src/providers/index.ts
@@ -1,1 +1,2 @@
 export * from './ConditionalReduxProvider';
+export * from './ProvideMetrics';